Authentication using passport jwt completed

+++ controllers/api/v1/posts_api.js
@@ -25,29 +25,22 @@ module.exports.destroy = async function(req,res){
     try{
         const post = await Post.findById(req.params.id);
 
-        //.id means converting the object id into string
-        // if(post.user == req.user.id){
-
-
+        // .id means converting the object id into string
+        if(post.user == req.user.id){
             // await Like.deleteMany({likeable: post._id, onModel: 'Post'});
             // await Like.deleteMany({_id: {$in:post.comments}});
-
-
             post.deleteOne();
-
             // we are deleting all the comments of the post also
            await Comment.deleteMany({ post:req.params.id});
-
-
-
            return res.json(200,{
             message: "Post and associated comments deleted successfully!"
            })
-        // }
-        // else{
-        //     req.flash("error","You cannot delete this post!");
-        //     return res.redirect('back');
-        // }
+        }
+        else{
+            return res.json(200,{
+                message: "You cannot delete this post!"
+               });
+        }
 
     }catch(err){
